Два новых хука для получения категорий и предметов
