+++ frontend/src/components/_reusable/subcomponents/items-list/items-list.tsx
@@ -1,21 +1,14 @@
-import React, { FC, useEffect } from 'react';
+import React, { FC } from 'react';
 import { Item } from '#reusable/subcomponents/items-list/item/item';
 import { useSelector } from 'react-redux';
-import { isItemsRequestPendingSelector, itemsListSelector } from '#redux/selectors';
-import { useAppDispatch } from '#redux/store';
-import { Operations } from '#redux/operations/operations';
+import { itemsListSelector } from '#redux/selectors';
 import { useShowItemsRequestError } from '#src/js/hooks/use-show-items-request-error';
+import { useLoadItemsList } from '#src/js/hooks/use-load-items-list';
 
 export const ItemsList: FC = () => {
-  const dispatch = useAppDispatch();
   const itemsList = useSelector(itemsListSelector);
-  const isPending = useSelector(isItemsRequestPendingSelector);
 
-  useEffect(() => {
-    dispatch(Operations.getItemsList());
-    // eslint-disable-next-line
-  }, []);
-
+  useLoadItemsList();
   useShowItemsRequestError(true);
 
   if (itemsList?.length === 0) {
@@ -32,7 +25,7 @@ export const ItemsList: FC = () => {
         <div key={item.id} className={`uk-width-1-1 uk-width-1-2@m uk-padding-small`}>
           <Item {...item} />
         </div>
-      )) ?? null }
+      )) ?? null}
     </div>
   );
 };
