+++ frontend/src/js/hooks/use-load-items-list.ts
@@ -1,0 +1,17 @@
+import { useAppDispatch } from '#redux/store';
+import { useSelector } from 'react-redux';
+import { isItemsRequestPendingSelector, itemsListSelector } from '#redux/selectors';
+import { useEffect } from 'react';
+import { Operations } from '#redux/operations/operations';
+
+export const useLoadItemsList = (): void => {
+  const dispatch = useAppDispatch();
+  const itemsList = useSelector(itemsListSelector);
+  const isPending = useSelector(isItemsRequestPendingSelector);
+
+  useEffect(() => {
+    if (!isPending && !itemsList)
+      dispatch(Operations.getItemsList());
+    // eslint-disable-next-line
+  }, []);
+};
